Update README.md
Simply Art
A site created for artists that will eventually allow artists to sign up and add images of their work along with options to connect with people that want to buy their work or just share their talents with other artists. This is an ongoing project and will be updated at a later date with more features.

The project
This is my capstone one project for my software engineering bootcamp with Springboard. Below are it's details.
Deployment
Simply Art is hosted at this Here's what it can do...
Shows a random image of an artists’ work and its creator
This provides a user with a new artist every time they open up the app
Allows a user to search for a piece of artwork based using a search query.
Eventually this will allow for other users to search for an artist's name along with their own webpage with their work.
Shows an image of a single artists about info with a link to their portfolio/website
This will eventually be in a list of artists by genre (fine art, digital, sculpting etc. 
Navigation
Here is the general process for visiting the site.
There is no sign in required. This feature will be established at a later date. As for now a user can go through and click the page links at the top of the page.
The home page with generate a random image of an artist with their name
The about page serves as a little bit of a background into why this app was created.  
The discover page is where a user can look for existing images by artists. This will display 1 image at a time.
The artists page will show a random artist and some of their basic information
The Images API
This project uses the Unsplash API for all information and images. Their site can be found here.  
Technologies & Tools Used
As a full stack website, there were quite a few that went into the making of Instime.
HTML
Jinja (templating)
CSS
Standard library
PostgreSQL
To be used at a later date (databases are created but not implemented)
Python
Flask (backend framework for Python)
Flask WTF (creating & validating forms)
WTForms Alchemy (making forms from FSQLA models)
Flask Bcrypt (encrypting passwords)
Dotenv (storing API keys)